Improve window resizing
With this change in place, we now behave better when the terminal window
is made more narrow.

Before, we drew the footer using spaces. When the window was made more
narrow, we would keep the same number of spaces, which would cause the
footer to wrap, and the rest of the window to scroll up.

With this change in place, we now draw the footer using
clear-to-end-of-line. This means that when the window is made more
narrow, there are no spaces that can wrap, and the footer will stay on
the same line, and nothing will scroll.

This only works when there are no other lines on the screen that are
long.