Fix hyperlink rendering bug
Before this change, if the last char of a rendered line had a hyperlink,
that hyperlink would continue to the next line, even if the next line
did not have a hyperlink.

With this change in place, we now remove the hyperlink attribute when
a line ends.

Discovered since riff now hyperlinks file names:
https://github.com/walles/riff